Delete unusable files. Moving filet to subdirs.
