+++ server/db/connectionmgr.cpp
