+++ server/db/sqlitemgr.cpp
