+++ server/db/sqlitemgr.h
