+++ server/db/sqlmgr.cpp
