+++ server/db/sqlmgr.h
