+++ server/network/networkmanager.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "qobject.h"
 #include "tcpserver.h"
-#include "sqlmgr.h"
+#include "db/sqlmgr.h"
 #include "users/usercontrol.h"
 #include "requestsmanager.h"
 
@@ -14,8 +14,10 @@ private:
     SQLMgr *_sqlMgr;
 
     void connectSlots();
+signals:
+    void newClientConnected(QTcpSocket *client);
 public slots:
-    void newClient( QTcpSocket* );
+    void newClient(QTcpSocket* client);
     void inputQuery( QString, QTcpSocket* );
 public:
     NetworkManager(SQLMgr &sqlMgr, const int port );
