--- server/tester.cpp
@@ -1,10 +1,0 @@
-#include "tester.h"
-			   
-Tester::Tester()
-{
-}
-
-
-Tester::~Tester()
-{
-}
