+++ server/users/accesscontrol.h
@@ -2,7 +2,7 @@
 #define ACCESSCONTROL_H
 
 #include "users/user.h"
-#include "sqlmgr.h"
+#include "db/sqlmgr.h"
 
 namespace Permissions {
     enum Permission{
